+++ frontend/app/src/components/TelegramLogin.vue
@@ -11,26 +11,26 @@ const telegram = ref();
 const script = document.createElement("script");
 const emit = defineEmits(["callback"]);
 const props = defineProps({
-    size: {
-        type: String,
-        default: () => "large"
-    },
-    onAuth: {
-        type: Function,
-        default: () => {}
-    },
-    botName: {
-        type: String,
-        default: () => "ETUTestingBot"
-    },
-    requestAccess: {
-        type: String,
-        default: () => "write"
-    }
+  size: {
+    type: String,
+    default: () => "large"
+  },
+  onAuth: {
+    type: Function,
+    default: () => {}
+  },
+  botName: {
+    type: String,
+    default: () => "ETUTestingBot"
+  },
+  requestAccess: {
+    type: String,
+    default: () => "write"
+  }
 });
 
 function onTelegramAuth(user: any) {
-    emit('callback', user);
+  emit('callback', user);
 }
 
 window.onTelegramAuth = onTelegramAuth;
@@ -43,10 +43,10 @@ script.setAttribute("data-request-access", props.requestAccess);
 script.setAttribute("data-onauth", "window.onTelegramAuth(user)");
 
 onMounted(() => {
-    telegram.value.appendChild(script)
+  telegram.value.appendChild(script)
 });
 </script>
 
 <template>
-    <div id="telegramAuth" ref="telegram"></div>
+  <div id="telegramAuth" ref="telegram"></div>
 </template>
