+++ frontend/app/src/entities/PagesEntity.ts
@@ -1,8 +1,8 @@
 export default interface Pages {
-    id: string;
-    name: string;
-    teacher_column: string;
-    columns: string[];
-    rule?: string;
-    notification_text?: string;
+  id: string;
+  name: string;
+  teacher_column: string;
+  columns: string[];
+  rule?: string;
+  notification_text?: string;
 }
