+++ frontend/app/src/entities/TableEntity.ts
@@ -1,10 +1,10 @@
 import type Pages from './PagesEntity';
 
 export default interface TableItem {
-	_id?: string;
-	name: string;
-	table_id: string;
-	provider: string;
-	update_frequency: number;
-	pages: Pages[];
+  _id?: string;
+  name: string;
+  table_id: string;
+  provider: string;
+  update_frequency: number;
+  pages: Pages[];
 }
