+++ frontend/app/src/entities/TeacherEntity.ts
@@ -1,8 +1,8 @@
 export default interface TeacherItem {
-    name: string;
-    patronymic: string;
-    surname: string;
-    telegram_login: string;
-    role: string;
-    _id?: string;
+  _id?: string;
+  name: string;
+  patronymic: string;
+  surname: string;
+  telegram_login: string;
+  role: string;
 }
