+++ frontend/app/src/models/TableModel.ts
@@ -2,51 +2,51 @@ import type TableItem from '@/entities/TableEntity';
 import type Pages from '@/entities/PagesEntity';
 
 export default class TableCreator {
-	data: TableItem[];
-	
-	constructor() {
-		this.data = [];
-	}
+  data: TableItem[];
 
-	addTable(table: TableItem) {
-		const exists = this.data.some(existingTable => existingTable._id === table._id);
-		if (!exists) {
-			this.data.unshift(table);
-		}
-	}
+  constructor() {
+    this.data = [];
+  }
 
-	addTableRule(table: TableItem, page: Pages) {
-		const existingTable = this.data.find(existingTable => existingTable._id === table._id);
-		if (existingTable) {
-			existingTable.pages.unshift(page);
-		}
-	}
+  addTable(table: TableItem) {
+    const exists = this.data.some(existingTable => existingTable._id === table._id);
+    if (!exists) {
+      this.data.unshift(table);
+    }
+  }
 
-	changeTableRule(page: Pages, tableId: String) {
-		const existingTable = this.data.find(existingTable => existingTable._id === tableId);
-		if (existingTable) {
-			const existingPage = existingTable.pages.find(existingPage => existingPage.id === page.id);
-			if (existingPage) {
-				existingPage.name = page.name;
-				existingPage.teacher_column = page.teacher_column;
-				existingPage.columns = page.columns;
-				existingPage.rule = page.rule;
-				existingPage.notification_text = page.notification_text;
-			}
-		}
-	}
+  addTableRule(table: TableItem, page: Pages) {
+    const existingTable = this.data.find(existingTable => existingTable._id === table._id);
+    if (existingTable) {
+      existingTable.pages.unshift(page);
+    }
+  }
 
-	changeTable(table: TableItem) {
-		const existingTable = this.data.find(existingTable => existingTable._id === table._id);
-		if (existingTable) {
-			existingTable.name = table.name;
-			existingTable.table_id = table.table_id;
-			existingTable.provider = table.provider;
-			existingTable.update_frequency = table.update_frequency
-		}
-	}
+  changeTableRule(page: Pages, tableId: String) {
+    const existingTable = this.data.find(existingTable => existingTable._id === tableId);
+    if (existingTable) {
+      const existingPage = existingTable.pages.find(existingPage => existingPage.id === page.id);
+      if (existingPage) {
+        existingPage.name = page.name;
+        existingPage.teacher_column = page.teacher_column;
+        existingPage.columns = page.columns;
+        existingPage.rule = page.rule;
+        existingPage.notification_text = page.notification_text;
+      }
+    }
+  }
 
-	removeTable(table: TableItem) {
-		this.data = this.data.filter(existingTable => existingTable._id !== table._id);
-	}
+  changeTable(table: TableItem) {
+    const existingTable = this.data.find(existingTable => existingTable._id === table._id);
+    if (existingTable) {
+      existingTable.name = table.name;
+      existingTable.table_id = table.table_id;
+      existingTable.provider = table.provider;
+      existingTable.update_frequency = table.update_frequency
+    }
+  }
+
+  removeTable(table: TableItem) {
+    this.data = this.data.filter(existingTable => existingTable._id !== table._id);
+  }
 }
