+++ frontend/app/src/views/AdminPanel.vue
@@ -4,12 +4,9 @@ import { RouterView } from 'vue-router';
 import RouteBar from "@/components/RouteBar.vue";
 import { useTablesStore } from '@/stores/tablesStore';
 import { useTeachersStore } from "@/stores/teachersStore";
-import axios from "@/config/defaultAxios";
-import type TeacherItem from "@/entities/TeacherEntity";
 
 const tablesStore = useTablesStore();
 const teachersStore = useTeachersStore();
-
 
 onMounted(() => {
   tablesStore.getTables();
